pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "season-me"
version = "0.1.0"
description = "Upload a selfie → personal color season analysis with a flattering 8-color palette."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "Pillow>=10.0",
    "numpy>=1.24",
    "opencv-python-headless>=4.8",
    "typer>=0.9",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
]

[project.scripts]
season-me = "season_me.cli:main"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
